requirements.txt
jieba
baidu-aip
flask
flask-restful
requests